Added functionality for specifying a directory for client files, renamed the server class, and reworked tests to avoid calling private methods.
